Up and Down navigation

+++ cd.c
@@ -32,13 +32,12 @@ int		search(char *str)
     return (FALSE);
 }*/
 
-
-
 char   *reading_input(t_assen *assen)
 {
 	char	*str;
 	int     r;
 	char    *tmp;
+	char    *temp;
 	t_history history;
 	t_assen *climb;
 	int fd;
@@ -48,6 +47,7 @@ char   *reading_input(t_assen *assen)
 
     str = NULL;
     tmp = NULL;
+    temp = NULL;
     ft_memset(&history, 0, sizeof (t_history));
     if (!(str = (char *) malloc( BUFFER + 1)))
         return (NULL);
@@ -70,16 +70,18 @@ char   *reading_input(t_assen *assen)
         {
             r = read(fd, str, BUFFER);
             str[r] = '\0';
-            ft_putstr(str);
-           /* int i = 0;
-            while (str[i])
-                printf("str[i] ==> %d\n", str[i++]);*/
-         /*  if (str[0] == 127)
+          if (str[0] == 127)
            {
-               tputs(history.delete_char, 0, int_put);
+               tputs(history.line_start, 0, int_put);
+               tputs(history.clear, 0, int_put);
+               prompt(g_in_signal);
                tmp = delete_char(tmp);
+               if (climb->next == NULL)
+                   temp = delete_char(temp);
                ft_putstr(tmp);
-           }*/
+           }
+          else
+              ft_putstr(str);
             if (memcmp(str, history.up_arrow, 3) == 0)
             {
                 tputs(history.line_start, 0, int_put);
@@ -104,10 +106,16 @@ char   *reading_input(t_assen *assen)
                     ft_putstr(climb->cmd);
                     tmp = ft_strdup(climb->cmd);
                 }
+                else if (climb->next == NULL)
+                {
+                    ft_putstr(temp);
+                    tmp = ft_strdup(temp);
+                }
             }
-            else if (memcmp(str, history.down_arrow, 3) != 0 && memcmp(str, history.up_arrow, 3) != 0)
+            if (memcmp(str, history.down_arrow, 3) != 0 && memcmp(str, history.up_arrow, 3) != 0 && str[0] != 127)
             {
                 tmp = extend_re(str, tmp);
+                temp = extend_re(str, temp);
                 str = (char *) malloc(sizeof(char) * BUFFER + 1);
                 if (str == NULL)
                     return (NULL);
@@ -120,11 +128,9 @@ char   *reading_input(t_assen *assen)
             }
         }
     }
-    printf("tmp ==> [%s]\n", tmp);
     if (tcsetattr(fd, TCSANOW, &save) == 1)
         exit (-1);
     free(str);
-    //tputs(history.key_e, 2, int_put);
     return (tmp);
 }
